pycli/requirements.txt
pandas==1.0.3
python-dotenv==0.12.0
pysftp==0.2.9
python-dotenv==0.12.0